embed/requirements_search_api.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.0.0
